+++ lib/logfile.dart
@@ -5,7 +5,7 @@ class LogFile {
 
   static final LogFile instance = LogFile._();
 
-  Future<void> init(String logDir, {String fileLeading = "LOG-FILE-", int maxFileCount = 10, int maxFileLength = 1024 * 1024 * 10}) async{
-    return await initLogger(logDir, fileLeading: fileLeading, maxFileCount: maxFileCount, maxFileLength: maxFileLength).whenComplete(() => i("LOG-FILE-INITIATED-SUCCSSFULLY"));
+  Future<void> init(String logDir, {String fileLeading = "-||-", int maxFileCount = 10, int maxFileLength = 1024 * 1024 * 10}) async{
+    return await initLogger(logDir, fileLeading: fileLeading, maxFileCount: maxFileCount, maxFileLength: maxFileLength);
   }
 }